Big commit:
Adding diagonal commands so the paddle can move in 8 directions. (or none).

Ball:
-Changed how the ball gets its velocities from calculations. This makes collision handling easier.
-Added in observer class for wall hit notifications.
-Added in friction math.
-Fixed math issue that was being caused by the point of reference not being in the center of the ball.

Collision Handler:

All classes that move and also participate in collisions have a copy of this class. They call it whenever they are about to move
to see if they make a collision or not and to help resolve it. All classes who participate register to this class.
I might optimize the math a little, but the main thing I wanted to accomplish is that at no time, no matter how fast things are moving will
a ball move through a paddle (unless the paddle pinches it against a wall but I allowed this behavior for now).

Input handler:

Uses command design pattern to return commands based on input. These commands can be resolved at any time and passed to any object which implements Icommand. This menas my input is bound to any one object. I plan to use this fact later.

ScoreBoard:

Listens for ball collisions against the walls and updates accordingly. Use of observer pattern here.